added ExecutorUtils.java; adjusted pom.xml; improved StreamUtils, MatcherUtils, ComparatorUtils

+++ src/main/java/org/omnaest/utils/MatcherUtils.java
@@ -275,7 +275,7 @@ public class MatcherUtils
                                                         .stream()
                                                         .sorted(ComparatorUtils.builder()
                                                                                .of(Map.Entry.class)
-                                                                               .with(entry -> entry.getKey())
+                                                                               .with(entry -> (Integer) entry.getKey())
                                                                                .reverse()
                                                                                .build())
                                                         .forEach(entry ->
